Modified the UgridChecker10 code so it would integrate with the compliance-checker command line tool.
Previous commit had the initialization of the UgridChecker10 object requiring a dataset, which broke the compliance checker. Now it leverages the BaseNCCheck.setup() method to assign the dataset object and create the dictionary of meshes.
Also had to add _cc_display_headers. Edited the checks.md for better organization and added an omitted check for volume_shape_type.
The ugrid.cdl includes a second mesh variable of dimension 2; 3D checks have yet to be implemented and tested.