netCDF4 method change
A conflict during the TravisCI build with an earlier PR prompted changing
of the netCDF4 requirement in the requirements file from v.1.4.2 to a more
relaxed >=v.1.4.0. Versions prior to 1.4.2 do not have a Variable.get_dims()
method, so in order to keep the code functioning we now get the dimensions
in a slightly more verbose way.